# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/usuarios/registro": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/usuarios/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "emailOrUsername": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/usuarios/perfil": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/usuarios/emailExists/{email}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/usuarios/usernameExists/{username}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/runas/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/glifos/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/armaduras/coraza": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/armaduras/guantes": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/armaduras/pantalones": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/armaduras/botas": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/espadas/acero": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/espadas/plata": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/plantillas/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                },
                "nombrePlantilla": {
                  "example": "any"
                },
                "idCoraza": {
                  "example": "any"
                },
                "idGlifo1": {
                  "example": "any"
                },
                "idGlifo2": {
                  "example": "any"
                },
                "idGlifo3": {
                  "example": "any"
                },
                "idGuantes": {
                  "example": "any"
                },
                "idGlifo4": {
                  "example": "any"
                },
                "idGlifo5": {
                  "example": "any"
                },
                "idGlifo6": {
                  "example": "any"
                },
                "idPantalones": {
                  "example": "any"
                },
                "idGlifo7": {
                  "example": "any"
                },
                "idGlifo8": {
                  "example": "any"
                },
                "idGlifo9": {
                  "example": "any"
                },
                "idBotas": {
                  "example": "any"
                },
                "idGlifo10": {
                  "example": "any"
                },
                "idGlifo11": {
                  "example": "any"
                },
                "idGlifo12": {
                  "example": "any"
                },
                "idEspadaAcero": {
                  "example": "any"
                },
                "idRuna1": {
                  "example": "any"
                },
                "idRuna2": {
                  "example": "any"
                },
                "idRuna3": {
                  "example": "any"
                },
                "idEspadaPlata": {
                  "example": "any"
                },
                "idRuna4": {
                  "example": "any"
                },
                "idRuna5": {
                  "example": "any"
                },
                "idRuna6": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/noticias/main/{categoria}/{pagina}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "categoria",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pagina",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/noticias/numPaginas/{categoria}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "categoria",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/noticias/merchandising/{pagina}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "pagina",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/noticias/numPaginasMerch": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}